Update of some things

+++ src/main/kotlin/emortal/holeymoley/util/PotionUtil.kt
@@ -5,5 +5,5 @@ import net.minestom.server.potion.Potion
 import net.minestom.server.potion.PotionEffect
 
 fun CustomPotionEffect.asPotion(): Potion {
-    return Potion(PotionEffect.fromId(id.toShort())!!, amplifier, duration, showParticles(), showIcon(), isAmbient)
+    return Potion(PotionEffect.fromId(id.toInt())!!, amplifier, duration, showParticles(), showIcon(), isAmbient)
 }
